feat: Add support for custom light and dark themes
This commit adds support for custom light and dark themes in the  class. It introduces two new optional parameters,  and , which allow the user to specify their own custom themes. If no custom themes are provided, the default themes ( and ) will be used.

The  class has also been updated to accept the  and  parameters in its constructor. This allows for the initialization of the  with the custom themes if provided, or with the default themes if not.

In addition, the  class in  has been modified to remove the explicit passing of the  and  parameters to the  method of the . This is because the  now handles the initialization of the themes internally.

These changes provide more flexibility and customization options for the themes used in the application.